unified components for directive

+++ src/app/update-repos.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, HostListener, Input } from '@angular/core';
+import { Directive, HostListener, Input, ElementRef } from '@angular/core';
 import { RepoService } from "../app/repo.service";
 import { Repo } from "../app/repo";
 
@@ -7,13 +7,19 @@ import { Repo } from "../app/repo";
 })
 export class UpdateReposDirective {
 
-  constructor(public repoService: RepoService) { }
+  constructor(public repoService: RepoService, private el: ElementRef) { }
 
   @Input('appUpdateRepos') repo: Repo;
 
-  @HostListener('click') onDelete() {
-    this.repoService.deleteRepo(this.repo);
-    this.repo.archived = false;
+  @HostListener('click') onClick() {
+if (this.el.nativeElement.className === 'repo__delete') {
+  this.repoService.deleteRepo(this.repo);
+  this.repo.archived = false;
+  
+} else if (this.el.nativeElement.className === 'repo__add') {
+  this.repoService.addRepo(this.repo);
+    this.repo.archived = true;
+}
   }
 
 }
